requirements/dev.txt
-r ../requirements.txt

coverage==4.4.2
django-coverage-plugin==1.5.0
django-test-plus==1.0.22
pytest-django==3.1.2
pytest-sugar==0.9.0
factory-boy==2.9.2
ipdb==0.10.3
django-extensions==1.9.9
django-debug-toolbar==1.9.1
GitPython==2.1.8
Markdown==2.6.11
tblib==1.3.2
mypy==0.560
flake8==3.5.0
